Refactoring of fetch api tests
